+++ examples/00_simple/Gruntfile.js
@@ -1,4 +1,3 @@
-
 var path = require('path');
 
 var stylesheetsDir = 'assets/stylesheets';
@@ -119,7 +118,7 @@ module.exports = function(grunt) {
   grunt.registerTask('compile', ['handlebars', 'browserify', 'stylus']);
 
   // Run the server and watch for file changes
-  grunt.registerTask('server', ['runNode', 'compile', 'watch']);
+  grunt.registerTask('server', ['compile', 'runNode', 'watch']);
 
   // Default task(s).
   grunt.registerTask('default', ['compile']);